requirements.txt
absl-py==0.11.0
aiohttp==3.7.3
aiohttp-cors==0.7.0
aioredis==1.3.1
astunparse==1.6.3
async-timeout==3.0.1
atari-py==0.2.6
attrs==20.3.0
blessings==1.7
#bsuite @ git+git://github.com/deepmind/bsuite.git@57920a63d24b0571678db36e0d8b213c40b16f85
cachetools==4.2.1
certifi==2020.12.5
chardet==3.0.4
chex==0.0.7
click==7.1.2
cloudpickle==1.6.0
colorama==0.4.4
colorful==0.5.4
contextvars==2.4
cycler==0.10.0
#dataclasses==0.8
decorator==4.4.2
descartes==1.1.0
dm-acme==0.2.0
dm-control==0.0.322773188
dm-env==1.3
dm-haiku==0.0.3
dm-reverb==0.2.0
dm-sonnet==2.0.0
dm-tree==0.1.5
filelock==3.0.12
flatbuffers==1.12
frozendict==1.2
future==0.18.2
gast==0.3.3
glfw==2.0.0
google-api-core==1.25.1
google-auth==1.25.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
googleapis-common-protos==1.52.0
gpustat==0.6.0
grpcio==1.32.0
gym==0.18.0
h5py==2.10.0
hiredis==1.1.0
idna==2.10
idna-ssl==1.1.0
imageio==2.9.0
immutables==0.14
importlib-metadata==3.4.0
jax==0.2.9
jaxlib @  https://storage.googleapis.com/jax-releases/nocuda/jaxlib-0.1.60-cp36-none-manylinux2010_x86_64.whl
joblib==1.0.0
jsonschema==3.2.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
labmaze==1.0.3
lxml==4.6.2
Markdown==3.3.3
matplotlib==3.3.4
mizani==0.7.2
msgpack==1.0.2
multidict==5.1.0
networkx==2.5
numpy==1.19.5
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
opencensus==0.7.12
opencensus-context==0.1.2
opencv-python==4.5.1.48
opt-einsum==3.3.0
optax==0.0.2
palettable==3.3.0
pandas==1.1.5
patsy==0.5.1
Pillow==7.2.0
plotnine==0.7.1
portpicker==1.3.1
prometheus-client==0.9.0
protobuf==3.14.0
psutil==5.8.0
py-spy==0.3.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyglet==1.5.0
PyOpenGL==3.1.5
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
ray==1.1.0
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
#rlax @ git+git://github.com/deepmind/rlax.git@890bd75283bd9aa9e240e1bcef9c3d8bf8346ad9
rsa==4.7
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
scikit-image==0.17.2
scikit-learn==0.24.1
scipy==1.5.4
seaborn==0.11.1
six==1.15.0
statsmodels==0.12.2
sysv-ipc==1.1.0
tabulate==0.8.7
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
tensorflow-probability==0.12.1
termcolor==1.1.0
threadpoolctl==2.1.0
tifffile==2020.9.3
toolz==0.11.1
tqdm==4.56.0
trfl==1.1.0
typing-extensions==3.7.4.3
urllib3==1.26.3
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.6.3
zipp==3.4.0
